Misc functions for models
The example on http://www.pyflux.com/bayesian-vector-autoregression/ uses code which hard codes for the example for which variables the priors should be adjusted.  Later the plot_parameters also hard code the parameters to plot.  

For the first case I generalized it to a function, which can be used for models with an arbitrary number of variables and lags. For the latter case I have developed a function which: given the variable's position determines which the indices of the latent lag variables of the model. Of course both functions could be included in an OO-style in the model class. Since they could also be useful for other pyflux models, Ross should deceide, whether and where to intergrate trem.